좋아요 기능 추가

+++ src/main/java/com/cos/photogramstart/repository/LikesRepository.java
@@ -5,13 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface LikesRepository extends JpaRepository<Likes, Long> {
+public interface LikesRepository extends JpaRepository<Likes, Integer> {
 
     @Modifying
     @Query(value = "INSERT INTO likes(imageId, userId, createDate) VALUES(:imageId, :principalId, now())", nativeQuery = true)
-    int mLikes(int imageId, Long principalId);
+    int mLikes(int imageId, int principalId);
 
     @Modifying
     @Query(value = "DELETE FROM likes WHERE imageId = :imageId AND userId = :principalId", nativeQuery = true)
-    int mUnLikes(int imageId, Long principalId);
+    int mUnLikes(int imageId, int principalId);
 }
